requirements.txt
streamlit
torch
torchvision
azureml-core
matplotlib
numpy
pandas